cse1_api/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.0
importlib-metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
Werkzeug==3.1.3
zipp>=3.20.0
dicttoxml==1.7.4
PyJWT==2.9.0
pytest==8.3.3
